LayoutTests/fast/js/kde/completion-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(), i is 2
PASS caught is true
PASS successfullyParsed is true

TEST COMPLETE
